backend/requirements.txt
Flask==3.1.1
gunicorn==20.1.0
psycopg==3.2.2
psycopg[binary]==3.2.2
Flask-SQLAlchemy==3.1.1
pytest==8.4.1
pytest-flask==1.3.0
flask_migrate==4.1.0
flask-cors==6.0.1
python-dotenv==1.1.1
flask_bcrypt==1.0.1
flask-jwt-extended==4.7.1
